Correct physicians page mobile formatting, organize to PhysicianComponents folder
